+++ src/main/java/StatusChecker.java
@@ -1,5 +1,3 @@
-package pool;
-
 import com.amazonaws.services.ec2.AmazonEC2Client;
 import com.amazonaws.services.ec2.model.*;
